Removed deprecated `EncodeLearningCredential` and `EncodePidInSdJwtVc` utilities. Introduced `MsoMdocMdlV1Claims` and `SdJwtVcPidClaims` for claim definitions. Simplified `IssueMobileDrivingLicence` by consolidating claim logic.
