+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/IssuerContext.kt
@@ -310,9 +310,7 @@ fun beans(
     //
     // Signed metadata signing key
     //
-    registerBean(lazyInit = true) {
-        accessCertificate()
-    }
+    registerBean(lazyInit = true) { accessCertificate() }
 
     //
     // Adapters (out ports)
@@ -830,227 +828,3 @@ fun beans(
         }
     }
 }
-
-private object IssuerFactory {
-    data class Ctx(
-        val clock: Clock,
-        val timeZone: TimeZone,
-        val env: Environment,
-        val credentialIssuerId: CredentialIssuerId,
-        val validateProof: ValidateProof,
-        val storeIssuedCredential: StoreIssuedCredential,
-        val allocateStatus: AllocateStatus,
-        val generateNotificationId: GenerateNotificationId,
-    )
-
-    context(ctx: Ctx)
-    fun mdl(issuerSigningKey: IssuerSigningKey): IssueMdoc<MobileDrivingLicence> {
-        val validity = ctx.env.duration("issuer.mdl.mso_mdoc.encoder.duration") ?: 31.days
-        val jwtProofsSupportedSigningAlgorithms =
-            ctx.env.readNonEmptySet(
-                "issuer.mdl.jwtProofs.supportedSigningAlgorithms",
-                JWSAlgorithm::parse,
-            )
-        val deviceBinding =
-            DeviceBinding.ts3(
-                jwtProofsSupportedSigningAlgorithms,
-                PreferredKeyStorageStatusPeriod(validity),
-            )
-        val mdlIssuerReusePolicy = credentialReusePolicy(ctx.env, "issuer.mdl")
-        val generateNotificationId =
-            ctx.generateNotificationId.takeIf {
-                ctx.env.getProperty<Boolean>("issuer.mdl.notifications.enabled") ?: true
-            }
-        return IssueMobileDrivingLicence(
-            clock = ctx.clock,
-            getAttestationAttributes = GetMobileDrivingLicenceDataMock(),
-            issuerSigningKey = issuerSigningKey,
-            deviceBinding = deviceBinding,
-            credentialReusePolicy = mdlIssuerReusePolicy,
-            validity = validity,
-            validateProof = ctx.validateProof,
-            generateNotificationId = generateNotificationId,
-            storeIssuedCredential = ctx.storeIssuedCredential,
-            allocateStatus = ctx.allocateStatus,
-        )
-    }
-
-    context(ctx: Ctx)
-    fun pidInMdoc(
-        issuerSigningKey: IssuerSigningKey,
-        getAttestationAttributes: GetAttestationAttributes<PidAttributes>,
-    ): IssueMdoc<PidAttributes> {
-        val duration = ctx.env.duration("issuer.pid.mso_mdoc.encoder.duration") ?: 31.days
-        val jwtProofsSupportedSigningAlgorithms =
-            ctx.env.readNonEmptySet(
-                "issuer.pid.mso_mdoc.jwtProofs.supportedSigningAlgorithms",
-                JWSAlgorithm::parse,
-            )
-        val pidMsoMdocReusePolicy = credentialReusePolicy(ctx.env, "issuer.pid.mso_mdoc")
-
-        return IssueMsoMdocPid(
-            clock = ctx.clock,
-            getAttestationAttributes = getAttestationAttributes,
-            issuerSigningKey = issuerSigningKey,
-            deviceBinding =
-                DeviceBinding.ts3(
-                    jwtProofsSupportedSigningAlgorithms,
-                    PreferredKeyStorageStatusPeriod(duration),
-                ),
-            credentialReusePolicy = pidMsoMdocReusePolicy,
-            validity = duration,
-            validateProof = ctx.validateProof,
-            generateNotificationId =
-                ctx.generateNotificationId.takeIf {
-                    ctx.env.getProperty<Boolean>("issuer.pid.mso_mdoc.notifications.enabled") ?: true
-                },
-            storeIssuedCredential = ctx.storeIssuedCredential,
-            allocateStatus = ctx.allocateStatus,
-        )
-    }
-
-    context(ctx: Ctx)
-    fun pidInSdJwtVc(
-        issuerSigningKey: IssuerSigningKey,
-        getAttestationAttributes: GetAttestationAttributes<PidAttributes>,
-    ): IssueSdJwtVcPid {
-        val expiresIn = ctx.env.duration("issuer.pid.sd_jwt_vc.duration") ?: 31.days
-        val notUseBefore = ctx.env.duration("issuer.pid.sd_jwt_vc.notUseBefore")
-        val pidSdJwtVcReusePolicy = credentialReusePolicy(ctx.env, "issuer.pid.sd_jwt_vc")
-
-        val digestsHashAlgorithm =
-            ctx.env.getProperty<HashAlgorithm>("issuer.pid.sd_jwt_vc.digests.hashAlgorithm") ?: HashAlgorithm.SHA_256
-        val jwtProofsSupportedSigningAlgorithms =
-            ctx.env.readNonEmptySet("issuer.pid.sd_jwt_vc.jwtProofs.supportedSigningAlgorithms", JWSAlgorithm::parse)
-
-        return IssueSdJwtVcPid(
-            clock = ctx.clock,
-            timeZone = ctx.timeZone,
-            getAttestationAttributes = getAttestationAttributes,
-            issuerSigningKey = issuerSigningKey,
-            credentialIssuerId = ctx.credentialIssuerId,
-            digestsHashAlgorithm = digestsHashAlgorithm,
-            deviceBinding =
-                DeviceBinding.ts3(
-                    jwtProofsSupportedSigningAlgorithms,
-                    PreferredKeyStorageStatusPeriod(expiresIn),
-                ),
-            credentialReusePolicy = pidSdJwtVcReusePolicy,
-            validity = expiresIn,
-            validateProof = ctx.validateProof,
-            generateNotificationId =
-                ctx.generateNotificationId.takeIf {
-                    ctx.env.getBoolean("issuer.pid.sd_jwt_vc.notifications.enabled") ?: true
-                },
-            storeIssuedCredential = ctx.storeIssuedCredential,
-            allocateStatus = ctx.allocateStatus,
-            calculateNotUseBefore = notUseBefore?.let { duration -> { iat -> iat + duration } },
-        )
-    }
-
-    context(ctx: Ctx)
-    fun learningCredentialInSdJwtVc(
-        issuerSigningKey: IssuerSigningKey,
-        getPidData: GetAttestationAttributes<PidAttributes>,
-    ): IssueLearningCredential {
-        val jwtProofsSupportedSigningAlgorithms =
-            ctx.env.readNonEmptySet(
-                "issuer.learningCredential.jwtProofs.supportedSigningAlgorithms",
-                JWSAlgorithm::parse,
-            )
-        val learningCredentialReusePolicy =
-            credentialReusePolicy(ctx.env, "issuer.learningCredential")
-        val validity = Duration.parse(ctx.env.getProperty("issuer.learningCredential.validity", "P31D"))
-        val digestHashAlgorithm =
-            ctx.env.getProperty<HashAlgorithm>(
-                "issuer.learningCredential.sdJwtVc.encoder.digests.hashAlgorithm",
-            ) ?: HashAlgorithm.SHA_256
-
-        val notificationsEnabled =
-            ctx.env.getProperty<Boolean>("issuer.learningCredential.notifications.enabled") ?: true
-
-        return IssueLearningCredential(
-            sdJwtVcSerialization = SdJwtVcSerialization.Compact,
-            clock = ctx.clock,
-            getAttestationAttributes =
-                IssueLearningCredential.randomLearningCredentials(
-                    ctx.clock,
-                    getPidData,
-                ),
-            issuerSigningKey = issuerSigningKey,
-            digestsHashAlgorithm = digestHashAlgorithm,
-            deviceBinding =
-                DeviceBinding.Required(
-                    jwtProofsSupportedSigningAlgorithms,
-                    KeyAttestationRequirement.ts3(
-                        PreferredKeyStorageStatusPeriod(validity),
-                    ),
-                ),
-            credentialReusePolicy = learningCredentialReusePolicy,
-            validity = validity,
-            validateProof = ctx.validateProof,
-            generateNotificationId = ctx.generateNotificationId.takeIf { notificationsEnabled },
-            storeIssuedCredential = ctx.storeIssuedCredential,
-        )
-    }
-}
-
-private fun getPidDataFromKeyCloak(
-    clock: Clock,
-    timeZone: TimeZone,
-    env: Environment,
-    webClient: WebClient,
-): GetAttestationAttributes<PidAttributes> {
-    val keycloakProperties =
-        KeycloakConfigurationProperties(
-            env.getRequiredProperty<URL>("issuer.keycloak.server-url"),
-            env.getRequiredProperty("issuer.keycloak.authentication-realm"),
-            env.getRequiredProperty("issuer.keycloak.client-id"),
-            env.getRequiredProperty("issuer.keycloak.username"),
-            env.getRequiredProperty("issuer.keycloak.password"),
-            env.getRequiredProperty("issuer.keycloak.user-realm"),
-        )
-    return GetPidDataFromKeyCloak(
-        issuerCountry = env.getRequiredProperty("issuer.pid.issuingCountry").let(::IsoCountry),
-        issuingJurisdiction = env.getProperty("issuer.pid.issuingJurisdiction"),
-        clock = clock,
-        timeZone = timeZone,
-        webClient = webClient,
-        keyCloak = Url(keycloakProperties.serverUrl.toExternalForm()),
-        administrationClient =
-            AdministrationClient(
-                realm = Realm(keycloakProperties.authenticationRealm),
-                client = Credentials(username = keycloakProperties.clientId, password = null),
-                admin =
-                    Credentials(
-                        username = keycloakProperties.username,
-                        password = keycloakProperties.password,
-                    ),
-            ),
-        users = Realm(keycloakProperties.userRealm),
-    )
-}
-
-private fun validateProof(
-    credentialIssuerId: CredentialIssuerId,
-    isTrustedKeyAttestationIssuer: IsTrustedKeyAttestationIssuer,
-    verifyNonce: VerifyNonce,
-): ValidateProof {
-    val verifyKeyAttestation = VerifyKeyAttestation(isTrustedKeyAttestationIssuer = isTrustedKeyAttestationIssuer)
-    val validateJwtProofWithKeyAttestation =
-        ValidateJwtProofWithKeyAttestation(credentialIssuerId, verifyKeyAttestation)
-    val validateAttestationProof = ValidateAttestationProof(verifyKeyAttestation)
-    return ValidateProof(
-        validateJwtProofWithKeyAttestation,
-        validateAttestationProof,
-        verifyNonce,
-    )
-}
-
-private fun httpProxy(env: Environment): HttpProxy? =
-    env.getProperty("issuer.http.proxy.url")?.let {
-        val url = Url(it)
-        val username = env.getProperty("issuer.http.proxy.username")
-        val password = env.getProperty("issuer.http.proxy.password")
-        HttpProxy(url, username, password)
-    }
